+++ src/services/primitive/apiClient.ts
@@ -1,4 +1,4 @@
-import axios from "axios";
+import axios from 'axios';
 
 export const apiClient = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_URL,
